# remove a typeunion in an extrapolation perform step method