Maybe a minor performance improvement.
Before
```julia
julia> @benchmark solve($prob, ImplicitHairerWannerExtrapolation(threading = true))
BenchmarkTools.Trial: 4239 samples with 1 evaluation.
 Range (min … max):  673.277 μs … 49.720 ms  ┊ GC (min … max):  0.00% … 98.15%
 Time  (median):       1.130 ms              ┊ GC (median):     0.00%
 Time  (mean ± σ):     1.175 ms ±  2.879 ms  ┊ GC (mean ± σ):  14.55% ±  5.81%

     ▁▁                                    ▄█▆▂
  ▂▃▆███▅▃▂▂▂▁▁▂▁▂▁▁▁▁▂▁▁▂▁▂▂▂▂▂▁▁▂▁▁▁▁▁▂▄▇████▆▃▂▂▂▂▂▂▁▂▂▂▂▁▂ ▃
  673 μs          Histogram: frequency by time         1.33 ms <

 Memory estimate: 559.05 KiB, allocs estimate: 5978.
```
After
```julia
julia> @benchmark solve($prob, ImplicitHairerWannerExtrapolation(threading = true))
BenchmarkTools.Trial: 6086 samples with 1 evaluation.
 Range (min … max):  439.537 μs … 37.903 ms  ┊ GC (min … max):  0.00% … 97.49%
 Time  (median):     825.197 μs              ┊ GC (median):     0.00%
 Time  (mean ± σ):   817.550 μs ±  2.276 ms  ┊ GC (mean ± σ):  17.04% ±  6.01%

     █▁                                                 ▆▆
  ▂▃▆██▅▃▂▂▂▁▁▁▁▂▁▁▁▂▁▁▁▁▂▂▁▂▂▂▂▂▂▁▁▁▁▂▁▁▁▁▁▁▂▁▂▂▁▂▁▁▂▃▇███▄▃▂ ▃
  440 μs          Histogram: frequency by time          869 μs <

 Memory estimate: 553.92 KiB, allocs estimate: 5794.
```